Add files via upload
Overview
Business Need
You work at Nexthikes as a Machine Learning Engineer and a company called Rossmann
Pharmaceuticals has given you a project on sales forecasting. The finance team wants to
forecast sales in all their stores across several cities six weeks ahead of time. Managers in
individual stores rely on their years of experience as well as their personal judgement to
forecast sales.
The data team identified factors such as promotions, competition, school and state
holidays, seasonality, and locality as necessary for predicting sales across the various
stores.
Your job is to build and serve an end-to-end product that delivers this prediction to
analysts in the finance team.